Bootstrap `metaplex-indexer-graphql`.
