--- crates/graph/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "metaplex-graph-server"
-version = "0.1.0"
-authors = ["Holaplex Engineering <hola@holaplex.com>"]
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-actix-web = "3"
-actix-cors = "0.4.0"
-env_logger = "0.8"
-serde = "1.0.103"
-serde_json = "1.0.44"
-serde_derive = "1.0.103"
-juniper = "0.15"
