+++ crates/rpc/Cargo.toml
@@ -5,12 +5,12 @@ authors = [
   "ryans <ryans@holaplex.com>",
 ]
 edition = "2021"
-description = "JSONRPC read replica for providing cached Metaplex data"
+description = "JSONRPC HTTP server for providing cached Metaplex data"
 documentation = "https://docs.rs/metaplex-indexer-rpc"
 readme = "../../README.md"
 repository = "https://github.com/holaplex/metaplex-indexer"
 license = "AGPL-3.0-or-later"
-keywords = ["solana", "metaplex", "holaplex", "web3"]
+keywords = ["solana", "metaplex", "holaplex", "web3", "jsonrpc"]
 categories = ["cryptography::cryptocurrencies", "web-programming"]
 
 [dependencies]
